django/19/07/requirements.txt
django==5.0.6
Pillow==10.3.0

djangorestframework==3.15.0
drf-spectacular==0.27.1